+++ RedPandaIDE/editor.h
@@ -316,7 +316,7 @@ private:
     TipType getTipType(QPoint point, QSynedit::BufferCoord& pos);
     void cancelHint();
     QString getHeaderFileHint(const QString& s, bool fromNext);
-    QString getParserHint(const QStringList& expression,int line);
+    QString getParserHint(const QStringList& expression,int line, int ch);
     void showDebugHint(const QString& s,int line);
     QString getErrorHint(const PSyntaxIssue& issue);
     QString getHintForFunction(const PStatement& statement,
@@ -336,6 +336,8 @@ private:
 
     Editor* openFileInContext(const QString& filename);
     bool needReparse();
+
+    PStatement constructorToClass(PStatement constuctorStatement, int line, int ch);
 
 private:
     bool mInited;