+++ NEWS.md
@@ -15,6 +15,8 @@ Red Panda C++ Version 3.4
   - enhancement: Show modified sign in problem set's title. 
   - fix: When hovering on #include header names, tips doesn't show.
   - fix: Class name is wrongly syntaxed colored in its definition when it has a constructor.
+  - fix: Mouse tip of class displays its constructor info in its definition.
+  - fix: Goto defintion/declaration will jump to constructor instead of the class in the class definition.
   
 Red Panda C++ Version 3.3
   - enhancement: Auto indent contents in parenthesis.